- Slipt peers directive file into 2 separated directives. - Basics upgradation to ES6
