+++ src/shared/peers/index.js
@@ -1,0 +1,2 @@
+import './peers.directive';
+import './os-icon.directive';
